Cycles moved to a separate file, removed old files
