+++ fetch_cord/computer/cpu/Cpu_intel.py
@@ -1,3 +1,4 @@
+from __future__ import annotations
 from .Cpu_interface import Cpu_interface
 import re
