+++ fetch_cord/computer/cpu/Cpu_interface.py
@@ -1,3 +1,4 @@
+from __future__ import annotations
 from abc import ABCMeta, abstractmethod
 from typing import Dict
