+++ fetch_cord/computer/cpu/get_cpu.py
@@ -1,3 +1,4 @@
+from __future__ import annotations
 from typing import List
 
 from .Cpu_amd import Cpu_amd
