+++ fetch_cord/computer/gpu/Gpu_interface.py
@@ -1,3 +1,4 @@
+from __future__ import annotations
 from abc import ABCMeta, abstractmethod
 from typing import List, TypeVar, Dict
