+++ fetch_cord/computer/gpu/Gpu_nvidia.py
@@ -1,10 +1,13 @@
+from __future__ import annotations
+from fetch_cord.run_command import BashError, exec_bash
 from .Gpu_interface import Gpu_interface
-import re
 
 GPU_VENDOR = "nvidia"
 
 
 class Gpu_nvidia(Gpu_interface):
+    primeoffload: bool
+
     def __init__(self, os, model):
         super().__init__(os, GPU_VENDOR, model)
 
@@ -22,8 +25,20 @@ class Gpu_nvidia(Gpu_interface):
                 "Temperature report for Nvidia GPU's is not supported on MacOS yet."
             )
         elif self.os == "linux":
-            raise NotImplementedError(
-                "Temperature report for Nvidia GPU's is not supported on Linux yet."
-            )
+            try:
+                return exec_bash(
+                    "nvidia-smi -q | awk '/GPU Current Temp/{print $5}' | sed 's/^/[/;s/$/°C]/'"
+                )
+            except BashError:
+                pass
         else:
             raise NotImplementedError("Unknown OS, no GPU temperature report.")
+
+    def check_primeoffload(self):
+        # only show the GPU in use with optimus, show both if prime render offload
+        self.primeoffload = False
+        try:
+            self.primeoffload = exec_bash('xrandr --listproviders | grep -o "NVIDIA-0"')
+            return True
+        except BashError:
+            return False
