+++ fetch_cord/computer/gpu/get_gpu.py
@@ -1,3 +1,4 @@
+from __future__ import annotations
 from typing import List
 
 from .Gpu_amd import Gpu_amd
