--- fetch_cord/ressources/infos.json
@@ -1,178 +1,0 @@
-{
-    "amdcpus": {
-        "ryzen 3": "741153175779803146",
-        "ryzen 5": "741152732756312125",
-        "ryzen 7": "740752899054895105",
-        "ryzen 9": "741152930899427364",
-        "ryzen threadripper": "742075019257184338"
-    },
-    "intelcpus": {
-        "intel i3": "741044208512532570",
-        "intel i5": "741099939198926920",
-        "intel i7": "741100300219187335",
-        "intel i9": "741100622040006719",
-        "intel pentium": "741203845706940467",
-        "intel celeron": "742904581360713849",
-        "pentium": "741203845706940467",
-        "intel core 2 duo": "746527640382603375",
-        "intel xeon": "749011141090607144"
-    },
-    "gpus": [
-        "intel",
-        "nvidia",
-        "amd",
-        "vmware",
-        "virtio",
-        "cirrus"
-    ],
-    "multigpus": {
-	"amdnvidia": "nvidiaamd",
-	"nvidiaamd": "nvidiaamd",
-	"intelnvidia": "nvidiaintel",
-	"nvidiaintel": "nvidiaintel",
-	"intelamd": "amdintel",
-	"amdintel": "amdintel"
-    },
-    "distros": {
-        "ubuntu": "740434138036699178",
-        "opensuseleap": "740156532137787433",
-        "arch": "740476198437650473",
-        "arco": "745435867971321866",
-        "artix": "741918141248045107",
-        "fedora": "740485660703719464",
-        "void": "740484961353597039",
-        "gentoo/linux": "740484380652208140",
-        "funtoo": "740484380652208140",
-        "centos": "740483295388631071",
-        "debian": "740490017218232392",
-        "opensusetumbleweed": "742180413132505088",
-        "manjaro": "740614258177605642",
-        "linuxmint": "740633577481568317",
-        "lmde": "741726946588622988",
-        "pop!_os": "740660055925587978",
-        "endeavouros": "740809641545564170",
-        "windows10": "741949889465942099",
-        "windows7": "741952383512346696",
-        "windows8": "741952179488948324",
-        "windows8.1": "741952065294827520",
-        "nixos": "744644133494325329",
-        "instantos": "744784599653285938",
-        "freebsd": "745054697047457822",
-        "solus": "745689115944681683",
-        "zorin": "745689865798156379",
-        "amazon": "749014983920320583",
-        "bedrock": "749015250904416287"
-    },
-    "versions": {
-        "10.13": "hsierria",
-        "10.14": "mojave",
-        "10.15": "catalina"
-    },
-    "windowmanagers": [
-        "dwm",
-        "i3",
-        "awesome",
-        "enlightenment",
-        "bspwm",
-        "xmonad",
-        "sway",
-        "openbox",
-        "herbstluftwm"
-    ],
-    "desktops": [
-        "kde",
-        "plasma",
-        "xfce",
-        "budgie",
-        "gnome",
-        "deepin",
-        "cinnamon",
-        "mate",
-        "aero",
-        "pantheon",
-        "explorer"
-    ],
-    "terminals": {
-        "st": "741280043220861030",
-        "kitty": "741285676250824725",
-        "alacritty": "741291339945345045",
-        "xterm": "741287143187546125",
-        "konsole": "741286819676553258",
-        "dolphin": "741286819676553258",
-        "gnome-terminal": "741328861115056160",
-        "cool-retro-term": "741731097498353794",
-        "urxvt": "743246048968835092",
-        "xfce4-terminal": "744332423072055296",
-        "apple_terminal": "744950796298354689",
-        "lxterminal": "745701997503840380",
-        "yakuake": "741286819676553258",
-        "mate-terminal": "746531413360377856"
-    },
-    "shells": [
-        "fish",
-        "zsh",
-        "bash",
-        "tcsh",
-        "ruby"
-    ],
-    "hosts": {
-        "aorus": "749061181607772200",
-        "inspiron": "743970870631858288",
-        "latitude": "743970870631858288",
-        "g3": "743970870631858288",
-        "hp": "743971270395297852",
-        "tuf": "744330890343219262",
-        "asus": "743936082780880928",
-        "asustek": "743936082780880928",
-        "acer": "744326890512318544",
-        "thinkpad": "744326223412461630",
-        "lenovo": "744326223412461630",
-        "gigabyte": "746447043190456561",
-        "macbookair,": "748290608787095694",
-        "macbookpro,": "750016315854553138",
-        "aspire": "744326890512318544",
-        "hvm": "749014556604497972"
-    },
-    "motherboards": {
-        "inspiron": "dell",
-        "latitude": "dell",
-        "g3": "dell",
-        "hp": "hp",
-        "tuf": "tuf",
-        "asus": "asus",
-        "asustek": "asus",
-        "acer": "acer",
-        "thinkpad": "lenovo",
-        "lenovo": "lenovo",
-        "gigabyte": "gigabyte"
-    },
-    "hostlist": [
-        "Acer",
-        "AORUS",
-        "Aspire",
-        "TUF",
-        "HP",
-        "ThinkPad",
-        "Inspiron",
-        "Lenovo",
-        "Latitude",
-        "G3",
-        "Asus",
-        "ASUSTeK",
-        "Gigabyte",
-        "HVM"
-    ],
-    "terminallist": [
-        "st",
-        "kitty",
-        "alacritty",
-        "xterm",
-        "konsole",
-        "gnome-terminal",
-        "cool-retro-term",
-        "urxvt",
-        "xfce4-terminal",
-        "terminator",
-        "tmux"
-    ]
-}
