+++ fetch_cord/update.py
@@ -1,12 +1,9 @@
-import urllib.request
-import sys
-import os
+from __future__ import annotations
 
+import urllib.request, sys, os
 
 def update():
     print("Updating database...")
-    url = "https://raw.githubusercontent.com/MrPotatoBobx/FetchCord/master/fetch_cord/ressources/infos.json"
-    urllib.request.urlretrieve(
-        url, os.path.dirname(__file__) + "/ressources/infos.json"
-    )
+    url = "https://raw.githubusercontent.com/MrPotatoBobx/FetchCord/master/fetch_cord/resources/fetchcord_ids.json"
+    urllib.request.urlretrieve(url, os.path.dirname(__file__) + "/resources/fetchcord_ids.json")
     sys.exit(0)
